Updated the logic for 2 different conditions for sick leave that malar mentioned i.e. server side validation for file needed for greater than 2 days and leave approval logic for sick leaves if leave is less than equal to 2 days

+++ HRManagement/DTOs/LeavesDTOs/LeaveBalanceDto.cs
@@ -5,8 +5,8 @@
         public int LeaveTypeId { get; set; }
         public string LeaveTypeName { get; set; }
         public int DefaultAnnualAllocation { get; set; }
-        public int Used { get; set; }
-        public int Remaining { get; set; }
+        public decimal Used { get; set; }
+        public decimal Remaining { get; set; }
     }
 
 }
